requirements.txt
requests==2.31.0
eth_utils==2.2.0
pycryptodome==3.18.0
web3==6.8.0
python-decouple==3.8